migrate react entities to writeFiles/transform
