+++ generators/generator-base.js
@@ -1376,6 +1376,7 @@ class JHipsterBaseGenerator extends PrivateBase {
   }
 
   /**
+   * @deprecated
    * Copy JSX templates after stripping translation keys when translation is disabled.
    *
    * @param {string} source - path of the source file to copy from
